pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "compares xml and json files"
repository = "https://github.com/tyranich/python-project-lvl2"
authors = ["tyranich <tyranich@mail.ru>"]
license = "MIT"
packages = [
    { include = "gendiff"},
]
[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.2"
coverage = "^6.3.2"
pytest-cov = "^3.0.0"
PyYAML = "^6.0"
pytest-lazy-fixture = "^0.6.3"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
